feat: add product api routes

+++ build.sh
@@ -3,6 +3,11 @@
 # 版本號文件
 VERSION_FILE="version.txt"
 
+# 讀取 .env 文件
+if [ -f ".env" ]; then
+    source .env
+fi
+
 # 檢查版本號文件是否存在，如果不存在則創建並設置初始版本號
 if [ ! -f "$VERSION_FILE" ]; then
     echo "1" > "$VERSION_FILE"
@@ -29,30 +34,33 @@ if [ $? -eq 0 ]; then
     # 處理先前的版本
     PREVIOUS_VERSION=$((VERSION - 1))
     if [ $PREVIOUS_VERSION -ge 1 ]; then
-        # 標籤並推送先前的版本到 GCP
-        docker tag my-node-app:$PREVIOUS_VERSION asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
-        docker push asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
+        # 判斷是否需要推送到 GCP
+        if [ "$PUSH_GCP" = "true" ]; then
+            # 標籤並推送先前的版本到 GCP
+            docker tag my-node-app:$PREVIOUS_VERSION asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
+            docker push asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
 
-        if [ $? -eq 0 ]; then
-            echo "Docker image my-node-app:$PREVIOUS_VERSION pushed to GCP successfully."
-        else
-            echo "Failed to push Docker image my-node-app:$PREVIOUS_VERSION to GCP."
-        fi
+            if [ $? -eq 0 ]; then
+                echo "Docker image my-node-app:$PREVIOUS_VERSION pushed to GCP successfully."
+            else
+                echo "Failed to push Docker image my-node-app:$PREVIOUS_VERSION to GCP."
+            fi
 
-        # 刪除先前的版本映像
-        docker rmi my-node-app:$PREVIOUS_VERSION
-        if [ $? -eq 0 ]; then
-            echo "Docker image my-node-app:$PREVIOUS_VERSION removed successfully."
-        else
-            echo "Failed to remove Docker image my-node-app:$PREVIOUS_VERSION."
-        fi
+            # 刪除先前的版本映像
+            docker rmi my-node-app:$PREVIOUS_VERSION
+            if [ $? -eq 0 ]; then
+                echo "Docker image my-node-app:$PREVIOUS_VERSION removed successfully."
+            else
+                echo "Failed to remove Docker image my-node-app:$PREVIOUS_VERSION."
+            fi
 
-        # 刪除先前的版本映像
-        docker rmi asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
-        if [ $? -eq 0 ]; then
-            echo "Docker image asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION removed successfully."
-        else
-            echo "Failed to remove Docker image asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION."
+            # 刪除先前的版本映像
+            docker rmi asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION
+            if [ $? -eq 0 ]; then
+                echo "Docker image asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION removed successfully."
+            else
+                echo "Failed to remove Docker image asia-east1-docker.pkg.dev/driven-slice-355713/my-api/my-api:$PREVIOUS_VERSION."
+            fi
         fi
     fi
 else
